+++ src/features/question/components/organisms/QuestionAnswer/index.jsx
@@ -14,6 +14,11 @@ import QuestionFeedback from '../QuestionEmotionForm';
 import StyledQuestionAnswer from './styles';
 import { ReactComponent as Share } from '../../../assets/Share.svg';
 import SaraMaraContext from '../../../../main/stores/SaraMaraContext';
+import useFeedback from '../../../hooks/useFeedback';
+import Form from '../../../../../components/molecules/Form';
+import Textarea from '../../../../../components/atoms/Textarea';
+import Toast from '../../../../../components/molecules/Toast';
+import StyledFeedback from '../QuestionEmotionForm/styles';
 
 const buttonStyle = {
   display: 'flex',
@@ -27,7 +32,7 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
   const [landingAnswer, setLandingAnswer] = useState('');
   const [count, setCount] = useState(0);
   const { setKeyword } = useContext(SaraMaraContext);
-
+  const { value: CScontent, onChange, submitCSFeedback, isFeedback, returnFeedback } = useFeedback(quesionId);
   useEffect(() => {
     setCompletedAnswer(answer);
   }, [answer]);
@@ -42,6 +47,7 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
       return result;
     });
   }, 50);
+
   const navigate = useNavigate('');
   const QuestionNavigator = (from) => {
     refreshForm();
@@ -51,6 +57,15 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
     setKeyword(QuestionFormData.ItemValue);
   }, [QuestionFormData.ItemValue]);
 
+  const firstChar = type.charAt(0).toUpperCase() + type.slice(1);
+
+  const checkDisabled = () => {
+    return !CScontent?.length;
+  };
+
+  const checkCSLenght = (e) => {
+    if (e.target?.value.length <= 500) onChange(e);
+  };
   return (
     <StyledQuestionAnswer>
       <section className="question-top">
@@ -76,6 +91,9 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
         <div style={{ position: 'relative' }}>
           <AnswerContent type={type} answer={landingAnswer} />
         </div>
+        <section className="answer-bottom">
+          <QuestionFeedback type={type} quesionId={quesionId} />
+        </section>
         <div className="button-wrap">
           <Button
             bg={type === 'sara' ? Theme.color.saraSecondary : Theme.color.maraSecondary}
@@ -120,9 +138,7 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
             </Title>
           </Button>
         </div>
-        <section className="answer-bottom">
-          <QuestionFeedback type={type} quesionId={quesionId} />
-        </section>
+
         <section className="share">
           <Text size="14px" label="공유하기" />
           <Share
@@ -134,6 +150,78 @@ export default function QuestionAnwser({ type, answer, QuestionFormData, quesion
             }}
           />
         </section>
+        <StyledFeedback>
+          <Form onSubmit={(e) => submitCSFeedback(e)} className={isFeedback ? 'cs-form is-feedback' : 'cs-form'}>
+            {!isFeedback ? (
+              <>
+                <Form.Input>
+                  <Textarea
+                    id="feedback"
+                    className="feedback-text"
+                    isAutoHeight
+                    disabled={isFeedback}
+                    value={CScontent}
+                    onChange={(e) => checkCSLenght(e)}
+                    h="94px"
+                    ph={
+                      isFeedback
+                        ? '의견이 제출 됐습니다'
+                        : `${firstChar}에게 바라는 점 하고싶은\n얘기가 있다면 작성해주세요`
+                    }
+                    rows={1}
+                  />
+                </Form.Input>
+                <div className="button-wrap">
+                  <Text
+                    size="12px"
+                    label={`${CScontent?.length ? CScontent?.length : 0}/500`}
+                    color="#ddd"
+                    bold="500"
+                  />
+                  <Form.Button>
+                    <Button
+                      type="submit"
+                      className={checkDisabled() ? 'disable-button' : ''}
+                      disabled={checkDisabled()}
+                      bg={type === 'sara' ? Theme.color.saraSecondary : Theme.color.maraSecondary}
+                      onClick={(e) => submitCSFeedback(e)}
+                    >
+                      <Text
+                        label="의견 보내기"
+                        color={type === 'sara' ? Theme.color.saraPrimary : Theme.color.maraPrimary}
+                        size="14px"
+                        bold="700"
+                      />
+                    </Button>
+                  </Form.Button>
+                </div>
+              </>
+            ) : (
+              <Toast>
+                <div className="image" />
+                <Text
+                  size="14px"
+                  style={{ marginTop: 'auto' }}
+                  bold="700"
+                  color={Theme.color.black}
+                  label={
+                    type === 'sara'
+                      ? `${type.charAt(0).toUpperCase() + type.slice(1)} 에게 의견을 보내줘서 고맙샤`
+                      : `${type.charAt(0).toUpperCase() + type.slice(1)} 에게 의견을 보내줘서 고맙다먀`
+                  }
+                />
+                <Text
+                  label="의견 다시보내기"
+                  style={{ marginBottom: 'auto', cursor: 'pointer' }}
+                  color={type === 'sara' ? Theme.color.saraPrimary : Theme.color.maraPrimary}
+                  size="12px"
+                  bold="700"
+                  onClick={returnFeedback}
+                />
+              </Toast>
+            )}
+          </Form>
+        </StyledFeedback>
       </section>
     </StyledQuestionAnswer>
   );
